Move NodeAddresses and DomainStateString under actuator
Both functions are actually determining domain state, not participaing
in changing the domain. Thus, they belong under actuator.
Plus, both are used outside of the libvirt client.
